[v15] Add troubleshooting info to Usage and Billing page (#48082)
* Add troubleshooting info to Usage and Billing page

Turn the section on validating usage data into the intro paragraph for a
troubleshooting section. We don't actually need to compare the usage
reporting system to audit events, and removing the comparison makes this
information concise enough for an introductory paragraph.

Add H2s for issues we have identified with usage reporting and billing
on self-hosted clusters.

* Fix grammar in the Usage/Billing guide