Edit the Agent service reference intro (#48135)
Closes #47256

Avoid the impression that there is a dedicated Teleport SSH Service
reference guide by editing the introduction to the Agent Service
Reference section.

Service-specific references are not very popular and date from a time
when Teleport users would focus on an "Access" product for a specific
type of resource. Rather than add an SSH Service reference, point users
to the general Teleport configuration reference.

The remaining service-specific reference guides are legacy
documentation. While there is a case to be made for merging them into
more popular documentation pages, that is outside the scope of this
change.