Add intro paragraphs to docs pages
Closes #44532

Ensure that all docs pages include an introductory paragraph, which
spells out the scope of a page and indicates to a reader whether the
page is appropriate for their needs.

In some pages, the "How it works" section functions like an introductory
paragraph. In these cases, turn the "How it works" section into an
introductory paragraph and add a new "How it works" section.

Other pages already include an introductory paragraph with the H2
"Introduction". Standardize the docs on not including a heading before
the introductory paragraph.