Machine ID: Fix `tshwrap` destination dir handling
Recent CLI handling changes broke destination dir resolution for
commands that use `tshwrap.GetDestinationDir()` to resolve
destinations from either the CLI or config file.

The old behavior depends on implicit creation of an identity service
if `--destination-dir` was provided on the CLI and no config file is
loaded. We now no longer make this assumption and prefer to only
resolve explicit configurations, so this tweaks
`tshwrap.GetDestinationDir()` to explicitly handle the CLI parameter
rather than hoping it appears in the generated config.

While this is legacy functionality, it's not slated for removal in
v17. Most users should still prefer to use the new app and database
tunnel services wherever possible.

Fixes #48107