Prevent panic in proxy diffing
#47561 incorrectly translated the logic on when to call the diff
function for resources. The previous logic only called the differ
_if_ the resource had already been seen before, which guaranteed
that both the old and new resource provided to the diff function
were not nil. However, after the conversion to the generic watcher
the diff function was called for new resources, resulting in a nil
value being provided for the old resource and caused a panic.
The previous behavior has been restored, and the ProxyWatcher test
has been updated to set a diff function to prevent regressions.