Extend resource access request validation checks (#48138)
In #46780 we put a cap on the total number of resources that can
be requested in a single request, which helps avoid situations
where a large access request can exceed resource size limits and
break request listing.

This change expands on the validations by also checking that the sum
of the lengths of all of the requested IDs stays below a reasonable
limit.