[v17] [kube] use local image for kubernetes integration tests (#48382)
* [kube] use local image for kubernetes integration tests

This PR replaces the use of the `nginx:latest` Docker image in Kubernetes integration tests with a custom-built image based on `alpine:3.20.3`. This custom image includes a shell for `kubectl exec` integration tests and a compiled binary to act as an HTTP server.

This change addresses recent issues where test workflows failed due to problems downloading the `nginx` image.

* add information about the image

* source files from alpine cdn

* copy files from alpine

* Update fixtures/alpine/README.md

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

* Update fixtures/alpine/README.md

Co-authored-by: Alan Parra <alan.parra@goteleport.com>

* source files from alpine cdn (again)

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>