Okta: fix access request deletion reconciliation race condition in HA setup (#5361)
In Auth Server HA setup, each Auth Server will run its own access request reconciler for Okta. They
will be competing and because of that all calls to `oktaClient.*OktaAssignment` must be checked for
(non)existence for the reconciliation loops to properly finish and reconciled AccessRequest's to be
properly (de)registered.

Example in `onCreate` reconciliation function:

	// If the assignment already exists, register it locally and move on.
	_, err = a.oktaClient.CreateOktaAssignment(ctx, assignment)
	if err != nil && !trace.IsAlreadyExists(err) {
		return trace.Wrap(err)
	}

The check for `IsNotFound` was missing in `onDelete` so if the OktaAssignment for the AccessRequest
was scheduled for cleanup (SetCleanupTime) by another Auth Server reconciler AND already cleared in
the backend, then `onDelete` was constantly failing for this AccessRequest and reconciliation
re-queued indefinitely.

A valid question is: why it works most of the time? This is probably because it takes a while for
the backend to delete OktaAssignment after calling SetCleanupTime(clock.Now()) and all HA
reconciler have enough time to get and update the OktaAssignment in the backend.

Source-commit: gravitational/teleport.e@93ed7f7a9d32ea72f8cb5ff275ceb489eebd5a50
Source-pull-requests: gravitational/teleport.e#5361