Prepare to remove go-oidc/oauth2 from config packages
Adds a OAuthClientCredentials type directly into the servicecfg
package so that the dependency on go-oidc/oauth2 can be removed.
The PluginOAuthProviders.Slack field has been deprecated, but left
in place so as not to break teleport.e and a new
PluginOAuthProviders.SlackCredentials field has been added to use
the new type. Both fields are being populated with the same data
from the file config at the moment. Once teleport.e has been updated
to consume the credentials from PluginOAuthProviders.SlackCredentials
the original field will be removed.