deploy-cloud.sh warns about image version mismatch (#5369)
The base image tag is the version that the script uses when deploying to
cloud tenants and setting the automatic upgrades channel in proxy
config.
It should warn when the version of the base image does not match the
version that was likely used to build teleport.

Source-commit: gravitational/teleport.e@cf532a70f73ec8b320bfbb4f83a62f0c6acbc441
Source-pull-requests: gravitational/teleport.e#5369