[sshkeys] analyse host users and ssh authorized_keys more frequently
If a user exists on a host but the `.ssh/authorized_keys` file does not, `fsnotify` won't monitor its creation and won't inform the agent when a new key is added to `.ssh/authorized_keys`.

This PR changes that and makes the system watch more frequently for the users and their `authorized_keys` files so that we can monitor authorized keys existence.

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>