Relocate OIDCClaimsToTraits (#5374)
The helper function was only ever used by lib/auth/oidc and caused
go-oidc/jose to appear in the dependency tree of anything importing
lib/services. This copies the function, and its corresponding test
inline in the oidc package so that it can be removed from lib/services.

Source-commit: gravitational/teleport.e@25b812911375524c0183f5ab75542cd91411b805
Source-pull-requests: gravitational/teleport.e#5374