Machine ID: Fix usage of "starts" and "configures" terminology (#48494)
This fixes an issue in tbot's CLI help text where all "configure"
commands claimed to start a bot when they were actually generating
a configuration. This was caused by commands being reused for both
start and configure cases and the help text was statically defined.

This fixes the issue by introducing a CommandMode enum to indicate to
a command at construction time whether it will be used for the "start"
or "configure" case, and implements a stringer to return the
appropriate term for each case.