Allow querying of Membership-only Access List grants
The exitsting Access List grant query combines the grants from
both the owner and and member-of Access Lists hierarchies. At present
the AWS Identity Center Integration does not consider owner grants
when provisoning account assignments into AWS, and so it needs a way
to query grants just from the Access List membership hierarchy.

This patch adds a `GetInheritedMemberGrants()` function to the
accesslist package to allow such a query.