Docs: add discovery_group to teleport.yaml/discovery_service examples (#48362)
* Docs: add discovery_group to teleport.yaml/discovery_service examples

When the discovery_group is not configured, teleport will log a warning
message saying that it is recommended.
Some configuration examples do not include it, so when users use that
example to implement their own variation, it will output a warning
message.
After seeing that warning users might wonder if there's anything wrong
with their `teleport.yaml`.
Instead, we add a discovery_group to all examples so that users don't
get that message.

* explain discovery group

* use partial instead