[v17] Restore S3 FIPS and custom endpoint capabilities (#48821)
* Restore S3 FIPS and custom endpoint capabilities

When migrating the S3 events handler to use aws-sdk-go-v2 applying
the FIPS settings and custom endpoint were inadvertently dropped.
This restores the functionality, while also adding tests to ensure
that they are always respected going forward.

A similar test was added to the dynamodb events handler as well to
prevent any regressions with FIPS settings there.

* chore: Bump AWS modules (#48322)

* chore: Bump api/ modules

* Tidy other modules

* chore: Bump AWS SDKv2 modules

* chore: Bump github.com/awslabs/amazon-ecr-credential-helper/ecr-login

* Tidy integrations/ modules

* Replace usages of s3types.LifecycleRuleFilterMemberPrefix

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>