Make `accesslist.GetAncestorsFor()` public (#48784)
The Identity Center user & group provisioning system needs to re-provision
Access Lists in response to changes in any of their enclosed member lists.
This patch makes the existing `GetAncestorsFor()` function public so the IC
SCIM provisioning system can walk the list of ancestors in any AccessList
change and trigger a re-provision if necessary, without having to
re-implement the ancestor search.