Simplify Lint (Docs) (#48845)
Currently, `Lint (Docs)` configures git submodules for
`gravitational/teleport` within a `gravitational/docs` clone based on
the value of `$GITHUB_REF` or `$GITHUB_HEAD_REF` so it can fetch the
submodules and build the docs.

This change checks out `gravitational/teleport` using `actions/checkout`
instead, then copies the `teleport` clone into the `content` directory
of the `docs` clone. This way, we don't need to extract a branch name,
and can perform multiple checks on the `teleport` clone, e.g., the Vale
linter when we add it to `Lint (Docs)`.