Sanitize aws service custom endpoints (#49004)
There is a slight behavior change in aws-sdk-go-v2 regarding custom
endpoints. The new sdk requires the endpoints to be a valid URI, and
the legacy sdk would allow a hostname only enpoint and automatically
apply a scheme. To prevent breaking existing configurations when
upgrading to v17, we now have to apply the same logic to endpoints
manually that the legacy sdk used to apply for us.

Closes https://github.com/gravitational/teleport/issues/48760.