Ensure principals that no longer meet predicate get de-provisioned (#5484)
* Ensure principals that no longer meet predicate get de-provisioned

When a principal changes such that it no longer matches the inclusion
predicate in the Provsioning System, it should be deleted from the
downstream system.

Prior to this change, the Provisionig System would check the inclusion
predicate and, if it failed, just ignore the the change as the principal
was out side the provisioner's set of interest. This failed to handle the
situation where a principal transitions from being included to excluded,
as the principal was just ignored.

This change makes the Provisioning System check for an existing provisioning
state record when the predicate fails, and de-provisioning the principal if
one is found.

Addresses: gravitational/teleport#48959

* Address review feedback

Source-commit: gravitational/teleport.e@6aa8e84190692de2346b0a56409888027fb780f9
Source-pull-requests: gravitational/teleport.e#5484