Ensure principals that no longer meet predicate are handled correctly (#5487)
* Ensure principals that no longer meet predicate are handled correctly

Prior to this change, the Identity Center event handling system would
check the inclusion predicate and, if it failed, just ignore the the
change as the principal was out side the IC integrations's set of
interest. This failed to handle the situation where a principal
transitions from being included to excluded, as the principal was just
ignored.

This behaviour would leave an obsolete Assignment State record in the
backend. If the principal subsequently meets their predicate again, the
SCIM Provisioning System would re-create the downstream User or Group,
which assigns the principal a new External ID.

Rather tha create a new Assignment State record, the Identity Center
integration would find the obsolete record - containing an obsolete
External ID - and assume that it was valid. At this point, all attempts
to create or delete Account Assignments on the principal are doomed
to fail.

This patch updates the predicate check to test for the condition where
a principal is moving out of the provsioned set, and deletes their
Assignment State record. The actual downstream deletion is handled by
the SCIM provisioning system, so this is all the action that needs to
be taken.

Addresses: gravitational/teleport#48960

* Apply review feedback

* Apply review feedback

Source-commit: gravitational/teleport.e@68f7992178a27f8662b3622ba56eaa2353cdb384
Source-pull-requests: gravitational/teleport.e#5487