Attempt to deflake TestDynamicClientReuse (#49078)
The test was only waiting for a subset of services to be ready,
proceeding with the test, and then closing the process. This caused
a few problems that contributed to the flakiness. First, not calling
`process.Wait` resulted in some services still being active and
writing to the data directory while the testing framework was cleaning
up the temp directory. Second, adding the Wait alone, would cause
deadlocks because the test did not wait for all services to be
initialized and ready before shutting down.

In addition to making both of th changes above, the test was also
modified to reduce the number of services being launched to slightly
speed up the test.

Closes #46958.