Improve error messaging when attempting to load the license file (#5508)
We were swallowing the actual cause of the error and displaying the
same message whether the file was missing, we were looking in the
wrong location, or if there was an error reading the contents.

The error message now includes the underlying error as well as the
path at which we attempted to read.

Closes gravitational/teleport#24954

Source-commit: gravitational/teleport.e@c6946ef98890c1e54b0766735a09fdcb7e28b24f
Source-pull-requests: gravitational/teleport.e#5508