build: Fix "make create-github-release" (#49111)
* build: Fix "make create-github-release"

Fix the `create-github-release` make target to use "go run" to run the
tool that creates the github release. This used to run a tool in this
repository that was pre-built, so the "go run" was not needed before.
Now the tool has moved to a separate repository, this is required.

* build: Remove single-use vars for simplicity

Remove the `CHANGELOG` and `RELEASE_NOTES_GEN` vars that contained the
name of a Go main package to run in favor of a straightforward "go run"
command.