Make TestRemoteUser more robust against slow trust establishment
The assertions made in the test did not allow for any delay in the
CAs being propagated in response to updating the trust relationship.
This caused failures in cases where the remote CA did no yet exist,
or did not contain the updated role mapping when the remote client
attempted requests.

Closes https://github.com/gravitational/teleport/issues/47628.