Link to the status page for cloud
For someone unfamiliar with the status page as being the source of truth
for scheduled upgrades, the "Teleport Cloud" section on the "Upcoming Releases" document can give a false impression that the next update will be further in the future than it really is.

Adding a link to the status page will help discoverability of the next
upcoming change to Cloud, even if it isn't a "key deliverable" for that
quarter.