Expose logs of child Teleport processes
Configures Teleport subprocesses to log with the same configuration
and output location as the parent process. Prior to this all logs
from the child process disappeared into the void. This makes trying
to debug issues with launching subprocesses  challenging since the
only observability into them were the exit code.

There is one notable exception: this currently only works if the
parent process is logging to stdout/stderr. Support for writing
logs to files/syslog are not yet supported.