[v17] Add ListDevicesByUser protos (#49328)
* Add ListDevicesByUser protos (#48993)

These support the upcoming service to allow users to see a list of their
own registered devices in the web UI account settings page.

* Remove DeviceView from ListDevicesByUser (#49209)

Originally, we included the `View` field in order to add the Owner
property to safeguard the returned details of a device ensure the owner
matched the caller. However, this ended up not being conditionally used
(we would never _not_ use it) so we can remove it from the request.