AWS OIDC Dashboard: Auto Discovery summary endpoint
This PR adds a new endpoint which returns the summary of an integration:
- EC2, RDS and EKS found and enrolled resources

It will be used to let the user know about what's the current state of
the integration.

PS: the number of deployed ECS agents is not yet filled, but will be as
soon as the other PR lands.