Allow absolute docs links in the changelog (#49201)
This is because the changelog needs to include absolute docs links in
order for our release process to run smoothly, since we need to sync
release notes to the changelog. To do so, add a message control
directive to the changelog.

This change removes the comment that disables the remark messaging
linter, since we have removed this custom linter in favor of Vale.