Enroll EKS Cluster from Discover: wait for access entry (#49474)
After running this flow a couple more times, we discovered that some
times it takes longer than 25s.
From multiple tests we found out that it can take up to 30s.
Instead of setting that as the wait time, we add a little more to ensure
we are more resilient.