deflake `TestListUnifiedResources_WithLogins` (#49614)
This PR tries to fix the flakiness of the `TestListUnifiedResources_WithLogins` test. The primary issue identified was that the role was first created and then updated via a separate update call. Since the role, by default, grants wildcard access to all labels, the resources were being returned after the role was created but before the cache had synchronized with the updated role event.

As a result, the system failed to return the expected logins, except for the SSH logins, which were created during the initial role creation event.

Close #45086

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>