Add Account Assignment IDs to IC Account & App Permission Set Lists (#49747)
* Add Account Assignment IDs to IC App Permission Set List

Account Assignment resources represent potential Identity Center
account assignments that will be created when a Teleport user is
granted access to them via an Access Request.

The UI needs to know which resources represent a given (Account,
Permission Set) pair when building an access request, so this change
allows Teleport to express this relationship in the enclosing
Idenitity Center and App resources, rather than have the UI try
and deduce it.

* doc