Adjust levels of Vale rules (#49661)
Backports #48850

Change the error levels of Vale style rules to pave the way for
requiring the Vale linter. Use the `warning` level for all rules except
for the step numbering rule. This mostly preserves the existing "check
prose style" GitHub Actions job behavior, in which Vale style rules are
optional. The one rule we require is to ensure that numbered `Step n/d`
headings follow the correct numbering scheme.