UserTasks for DiscoverEKS Issues - proto
This PR adds the proto messages required to start emitting UserTasks for
DiscoverEKS issues.

What are UserTasks?
UserTasks are work items created to let the user know of failures that
occured when trying to run Auto Discover processes.
When something fails it will create a UserTask with as much detail as
possible. The user is then notified of a new UserTask, and can act upon
it.