postrelease: prefer the Go version in go.mod's toolchain directive (#49891)
The go tool will automatically remove the toolchain directive if it
happens to be the same as the go directive. This means in order to
reliably get the version of Go that is used to build, we need to first
check the toolchain directive, and fall back to the go directive only
if the toolchain is not specified.

Additionally, the toolchain directive is prefixed with "go", so we
need to strip that to get consistent output.