Adds AccountAsignment refs and Labels to IC Accounts (#5605)
Adds references to the appropriate Account Assignment resources
to the Account permision set list on synchonization.

This allows the correct Account Assignmnt names to be plumbed through
all the way to the UI, so that the UI does not have to generate account
assignment IDs from the available Account and Permission Set and deal
with thigs like name normalization or escaping rules.

Source-commit: gravitational/teleport.e@47432dde2cb0e19adac8da615638f18c2fb90b53
Source-pull-requests: gravitational/teleport.e#5605