Hide review box if user cannot review access requests (#49937)
oss counterpart for
https://github.com/gravitational/teleport.e/pull/5628

This adds some testing to the view as well as the equivalent to the web
solution for Connect. Connect was missing the [recently
added](https://github.com/gravitational/teleport/pull/48536)
`ReviewRequests` field in the user ACL, so I added it here.

Because this is handled in the tsh code, we don't have to worry about
backward compatibility here for Connect right?