Wait for process to exit before terminating tests
This attempts to reduce the flakiness of tests caused by failing
to clean up the data dir:

```
 testing.go:1232: TempDir RemoveAll cleanup: unlinkat /tmp/TestMonitor3923075621/001: directory not empty
```

The root cause of the flakiness is that the tests are exiting
prior to the Teleport process terminating. All places that Close
was called now include a Wait call to ensure things have all exited
before returning from the test.