Improve TeleportClient host resolution
Update tsh to prefer using ResolveSSHTarget to resolve the target
if labels, search, or predicate expressions were supplied. In the
event that the RPC is not implemented, tsh falls back to the
previous behavior of calling ListUnifiedResources, with one slight
addition. In the event that resolution is ambiguous tsh now
consults the cluster networking config to determine the
ROUTE_TO_MOST_RECENT strategy is in place. If so, the returned
server with the expiry the farthest in the future will be used.