Handle duplicate AWS Account names (#5693)
* Handle duplicate AWS Account names

It turns out that AWS account names are not guaranteed to be unique, so any
Teleport resources that use the account name for their IDs are at risk of
name clashes.

This patch adds the guranteed-unique AWS Account ID to resource names where
appropriate, and uses the normal reconciliation system to migrate the
resources to the new names.

Some manual cleanup may be needed for Users and Access Lists assigned roles
with the old names.

* Test fixup

Source-commit: gravitational/teleport.e@6b1edd13a093b4813da6156868526301f18fa42b
Source-pull-requests: gravitational/teleport.e#5693