Fix Integration Button styles (#5702)
If a user could list integrations, and they could NOT create
integrations, AND integrations existed, this would cause the disabled
'enroll new integration' button to be smashed to the left of the header
(this is due to how its wrapped in an optional Hover Tooltip).

This PR just adds some space-between flex property to the header to
ensure the button stays on the right side. It also moves the
HoverTooltip position to the bottom because it looks better

Source-commit: gravitational/teleport.e@b01c887629cd831e4bb48d85e126aa21f12cfa77
Source-pull-requests: gravitational/teleport.e#5702