web: remove usages of WithAuthCookieAndCSRF (#5688)
There were only two usages:

1. GET /enterprise/plugins/callback/:type
2. POST /enterprise/plugin

For item 1: WithAuthCookieAndCSRF does not validate CSRF tokens
for GET requests, so there is no behavior change here - we were
never doing any CSRF validation here.

For item 2: we have two alternate forms of protection:

a. We are now explicitly setting our session cookies to
   SameSite=Lax (see gravitational/teleport#50097), so our session
   cookie won't be sent on cross-site POST requests.
b. We use WithAuth, which checks for the Teleport bearer token in
   the Authorization header.

This is part 1 of a series of changes that will remove the CSRF
token cookie.

Also remove the CSRF token check from public cloud-only endpoints.
This is safe because these endpoints validate that the Content-Type
header is application/json.

Source-commit: gravitational/teleport.e@ddff99f4ebf0e2d6591b1db865a316c76e0868fc
Source-pull-requests: gravitational/teleport.e#5688