Convert lib/multiplexer to use slog (#50156)
* Tweak loglimit.LogLimiter

This updates the log limiter to implement slog.Handler and perform
sampling in line rather than relying on a separate goroutine. The
output is now slightly different as we no longer indicate how many
duplicate messages were received, but the overall functionality
of sampling spammy log messages should be the same.

* Convert lib/multiplexer to use slog