RDP: put the username in the load balancing cookie (#50225)
This doesn't change the behavior of the RDP server in any way,
but load balancers sitting in front of the server may use the
cookie to alter how they route connections.

Note that this is the default behavior of IronRDP when using
username/password authentication, but we must explicitly enable
it because we use smart card authentication instead.