Refactor the standard role editor to optimize its performance
This change makes every role editor section render separately, not
affecting other sections. This is achieved by property memoization. In
order to do it, we switch from managing the standard editor state using
`useState()` to `useReducer()`. The state and action dispatcher are
created with a new `useStandardModel()` function.

Another change required to support memoization in the rendering phase is
memoizing partial validation results.